Create a first simple model based on Bag of Words technique
The first model used for the classification of plots into genres is a Naive Bayes classifier, that uses a bag-of-words representation as input. This model although, it doesn't achieve great performance, it reveals some particularities of the dataset (caused by the imbalance of the classes) and can be used as a baseline for evaluation of the next models.